Вернул прежнюю структуру проекта
