+++ templates/main.php
@@ -57,9 +57,6 @@
 
     <div class="popular__posts">
         <?php foreach ($posts as $key => $value):
-            $timeHelper = new TimeHelper();
-            $stringHelper = new StringHelper();
-
             $post_content = htmlspecialchars($value['content_text']);
             $post_type = htmlspecialchars($value['icon_type']);
             $post_title = htmlspecialchars($value['title']);
@@ -117,7 +114,7 @@
                        </div>
                     <?php else: ?>
                         <!--здесь содержимое карточки-->
-                        <p><?= $stringHelper->cropText($post_content) ?></p>
+                        <p><?= cropText($post_content) ?></p>
                     <?php endif; ?>
                 </div>
                 <footer class="post__footer">
@@ -131,10 +128,10 @@
                             <div class="post__info">
                                 <b class="post__author-name"><?= $user_name ?></b>
                                 <?php
-                                $post_date = $timeHelper->GetPostTime($key);
+                                $post_date = GetPostTime($key);
                                 ?>
                                 <time class="post__time" title="<?= $post_date->format('d.m.Y H:i') ?>"
-                                      datetime="<?= $post_date->format('Y-m-d H:i:s') ?>"><?= $timeHelper->GetDateRelativeFormat($post_date); ?></time>
+                                      datetime="<?= $post_date->format('Y-m-d H:i:s') ?>"><?= GetDateRelativeFormat($post_date); ?></time>
                             </div>
                         </a>
                     </div>
